// src/api/consult/consult.types.d.ts
export declare enum CallEndedTypeId {
    Dropped = 1,
    Consulted = 2,
    InCompleteConsult = 3,
    Cancelled = 4,
    Initiated = 5,
    Scheduled = 6,
    Missed = 7,
    Rescheduled = 8,
    Offline = 9
}
// ...
export type ClinicalCommonDetails = {
    items: ClinicalCommonDataDetails[];
    note: string;
};
export type LabTestDetails = {
    data: any;
    note: string;
};
export type Medication = {
    dosage: string;
    duration: string;
    frequency: string;
    medicineId: string;
    medicineName: string;
    notes: string;
    route: string;
    schedules: Schedule[];
    timing: string;
};
// ...
export type CreatePrescriptionBody = {
    labTests: LabTestDetails[];
    medications: Medication[];
    diagnoses: ClinicalCommonDetails | Object;
    symptoms: ClinicalCommonDetails | Object;
    allergies: ClinicalCommonDetails | Object;
    medPreFiles: string[];
};
export type CreatePrescriptionResponse = ConsultSuccessResponse<CreatePrescriptionData, null>;
export type CreatePrescriptionData = {
    prescription: string;
};
export type GetAllSymptomsProps = {
    client: AxiosInstance;
    params: GetAllSymptomsParams;
    options?: CommonQueryOptions;
};
export type GetAllSymptomsParams = {
    name?: string;
    take: number;
    page: number;
};
export type ClinicalCommonDataDetails = {
    id: string;
    name: string;
    createdAt: string;
    updatedAt: string;
};
// ...
export type ConfirmConsultationBody = CreatePrescriptionBody & {
    placeMedOrder: boolean;
    placeLabOrder: boolean;
    package_id: number;
};